temp

+++ packages/DApp/src/helpers/loadProtons.ts
@@ -14,7 +14,7 @@ message WakuFeature {
   string voter = 1;
   bytes  sntAmount = 2;
   string community = 3;
-  uint64 timestamp = 4;
+  uint32 timestamp = 4;
   string sign = 5;
 }
 `)
